src/pyproject.toml
[tool.poetry]
name = "memorysafedataaggregator"
version = "0.1.0"
description = "A tool for scraping, aggregating, and analyzing data for a research project assessing memory safe languages"
authors = ["Joshua Ferguson <joshua.ferguson.273@gmail.com>"]
license = "Apache"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.25.2"
jupyter = "^1.0.0"
numpy = "^1.25.2"
cwe2 = "^2.0.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
nvdlib = "^0.7.4"
html5lib = "^1.1"
pyexploitdb = "^0.2.9"
pyserde = "^0.12.3"
orjson = "^3.9.9"
pandas = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
